+++ Source/Core/Views/UICollectionView/FlowLayouts/MLBusinessLayoutAttributeAnimator.swift
@@ -1,0 +1,14 @@
+//
+//  MLBusinessLayoutAttributeAnimator.swift
+//  MLBusinessComponents
+//
+//  Created by Gaston Maspero on 08/11/2020.
+//
+
+import Foundation
+
+protocol MLBusinessLayoutAttributeAnimator {
+    var factor: CGFloat { get }
+    
+    func transform(attributes: UICollectionViewLayoutAttributes, with scale: CGFloat)
+}